+++ ors-api/src/test/java/org/heigit/ors/apitests/routing/ResultTest.java
@@ -52,6 +52,7 @@ import java.util.List;
 import static io.restassured.RestAssured.given;
 import static io.restassured.config.JsonConfig.jsonConfig;
 import static org.hamcrest.Matchers.*;
+import static org.heigit.ors.apitests.utils.CommonHeaders.jsonContent;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -4589,6 +4590,8 @@ class ResultTest extends ServiceTest {
                 .body("error.message", is("Cannot compile expression: in 'priority' entry,  invalid expression \"äöü this is not a valid condition expression.\""))
                 .statusCode(500);
     }
+
+
 
     @Test
     void testBarriersAccessPermit() {